Edit Login Screen

+++ lib/model/user_model.dart
@@ -26,7 +26,7 @@ class UserDataLogin {
   String phone;
   String image;
   int points;
-  double credit;
+  int credit;
   String token;
 
   UserDataLogin.fromJson(Map<String, dynamic> data) {
